requirements.txt
bidict==0.22.1
certifi==2022.12.7
idna==2.10
numpy==1.24.2
pyjsparser==2.7.1
PyJWT==2.6.0
PyMySQL==1.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.32.3
six==1.16.0
sqlalchemy==2.0.41
SQLAlchemy_Utils==0.41.2
ta==0.11.0
ujson==5.10.0
urllib3==1.26.14
websocket-client==1.5.1
websockets==8.1
pandas==2.2.3
asyncio==3.4.3
